+++ test/test-applications/integrations/TestApplication.SelectiveSampler/Plugins/MixedModeSamplingPlugin.cs
@@ -1,35 +1,36 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
+using OpenTelemetry.AutoInstrumentation.PluginApi.ContinuousProfiling;
+using OpenTelemetry.AutoInstrumentation.PluginApi.SelectiveSampling;
 using OpenTelemetry.Trace;
 
 namespace TestApplication.SelectiveSampler.Plugins;
 
 #pragma warning disable CA1515 // Consider making public types internal. Needed for AutoInstrumentation plugin loading.
-public class MixedModeSamplingPlugin
+public class MixedModeSamplingPlugin : BasePlugin, ISelectiveSamplerPlugin, IContinuousProfilerPlugin
 #pragma warning restore CA1515 // Consider making public types internal. Needed for AutoInstrumentation plugin loading.
 {
-#pragma warning disable CA1822 // Mark members as static. Needed for AutoInstrumentation plugin loading.
-    public Tuple<bool, uint, bool, uint, TimeSpan, TimeSpan, object> GetContinuousProfilerConfiguration()
+    public ContinuousProfilerConfiguration GetFirstContinuousProfilerConfiguration()
     {
-        var threadSamplingEnabled = true;
-        var threadSamplingInterval = 200u;
-        var allocationSamplingEnabled = false;
-        var maxMemorySamplesPerMinute = 200u;
-        var exportInterval = TimeSpan.FromMilliseconds(50);
-        var exportTimeout = TimeSpan.FromMilliseconds(5000);
-        object continuousProfilerExporter = JsonConsoleExporter.Instance;
-
-        return Tuple.Create(threadSamplingEnabled, threadSamplingInterval, allocationSamplingEnabled, maxMemorySamplesPerMinute, exportInterval, exportTimeout, continuousProfilerExporter);
+        return new ContinuousProfilerConfiguration()
+        {
+            ThreadSamplingEnabled = true,
+            ThreadSamplingInterval = 200u,
+            AllocationSamplingEnabled = false,
+            MaxMemorySamplesPerMinute = 200u,
+            ExportInterval = TimeSpan.FromMilliseconds(50),
+            ExportTimeout = TimeSpan.FromMilliseconds(5000),
+            Exporter = JsonConsoleExporter.Instance
+        };
     }
 
-    public Tuple<uint, TimeSpan, TimeSpan, object> GetSelectiveSamplingConfiguration()
+    public SelectiveSamplerConfiguration? GetFirstSelectiveSamplingConfiguration()
     {
         return SelectiveSamplerPluginHelper.GetTestConfiguration();
     }
 
-    public TracerProviderBuilder BeforeConfigureTracerProvider(TracerProviderBuilder builder)
-#pragma warning restore CA1822 // Mark members as static. Needed for AutoInstrumentation plugin loading.
+    public override TracerProviderBuilder BeforeConfigureTracerProvider(TracerProviderBuilder builder)
     {
         builder.AddProcessor<FrequentSamplingProcessor>();
         return builder;